requirements.txt
streamlit
requests
python-dotenv
groq
chromadb
sentence-transformers
bcrypt
pandas
matplotlib